Installation script for the turtlebot2 kobuki
The Turtlebot is currently not supported in melodic and needs to be
build per source. Most of it is copied from a install script found here https://github.com/gaunthan/Turtlebot2-On-Melodic/blob/master/install_basic.sh
Seems to use the kinetic version although there are already git branches
for melodic. Might be worth to look at.