Adjustements after review
Added keyboard teleoperation for the docker version of the tutorial.

Removed the docker-scripts from CMakelist since the user does not need
to install the tutorial if docker is used.

Disabled rviz for the demo. It should be started deliberately and
preferably on the hostmachine. Adjusted tutorial accordingly.
Therefore the extra version of the tutorial launch for docker is no
longer necessary.